+++ packages/builder/builder-shared/src/devServer.ts
@@ -63,8 +63,6 @@ export const getDevServerOptions = async ({
     },
     source: {
       alias: {},
-      define: {},
-      globalVars: {},
     },
     html: {},
     tools: {