+++ packages/builder/builder-shared/src/types/config/source.ts
@@ -1,10 +1,18 @@
 import type { Alias } from '@modern-js/utils';
 import type { BuilderTarget } from '../builder';
+import type { ModifyChainUtils } from '../hooks';
 import type { ChainedConfig, JSONValue } from '../utils';
 
 export type ModuleScopes = Array<string | RegExp>;
 
 export type MainFields = (string | string[])[];
+
+export type GlobalVars = Record<string, JSONValue>;
+
+export type ChainedGlobalVars = ChainedConfig<
+  GlobalVars,
+  Pick<ModifyChainUtils, 'env' | 'target'>
+>;
 
 export interface SharedSourceConfig {
   /**
@@ -31,7 +39,7 @@ export interface SharedSourceConfig {
   /**
    * Define global variables. It can replace expressions like `process.env.FOO` in your code after compile.
    */
-  globalVars?: Record<string, JSONValue>;
+  globalVars?: ChainedGlobalVars;
   /**
    * Whether to compile JavaScript code imported via Data URI.
    */
@@ -48,7 +56,7 @@ export interface SharedSourceConfig {
   /**
    * Simple transformImport configuration
    */
-  transformImport?: SharedTransformImport[];
+  transformImport?: false | SharedTransformImport[];
 }
 
 export type SharedTransformImport = {
@@ -56,10 +64,12 @@ export type SharedTransformImport = {
   libraryDirectory?: string;
   style?: string | boolean;
   styleLibraryDirectory?: string;
+  camelToDashComponentName?: boolean;
+  transformToDefaultImport?: boolean;
 };
 
 export interface NormalizedSharedSourceConfig extends SharedSourceConfig {
   preEntry: string[];
-  globalVars: Record<string, JSONValue>;
+  globalVars: ChainedGlobalVars;
   compileJsDataURI: boolean;
 }