+++ tests/integration/doc/fixtures/plugin/plugin.ts
@@ -1,0 +1,20 @@
+import path from 'path';
+import { DocPlugin } from '@modern-js/doc-tools';
+
+export function docPluginDemo(): DocPlugin {
+  return {
+    name: 'doc-plugin-demo',
+    addPages() {
+      return [
+        {
+          routePath: '/filepath-route',
+          filepath: path.join(__dirname, 'blog', 'index.md'),
+        },
+        {
+          routePath: '/content-route',
+          content: '# Demo2',
+        },
+      ];
+    },
+  };
+}