+++ packages/builder/plugin-image-compress/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @modern-js/builder-plugin-image-compress
 
+## 2.14.0
+
+### Patch Changes
+
+- Updated dependencies [4779152]
+- Updated dependencies [9321bef]
+- Updated dependencies [9b45c58]
+- Updated dependencies [52d0cb1]
+- Updated dependencies [60a81d0]
+- Updated dependencies [dacef96]
+- Updated dependencies [16399fd]
+  - @modern-js/utils@2.14.0
+
 ## 2.13.4
 
 ### Patch Changes