+++ packages/generator/generators/dependence-generator/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/dependence-generator
 
+## 3.1.8
+
 ## 3.1.7
 
 ## 3.1.6